Fix a warning. Treated as error with default settings.

+++ src/XmlBlob.cpp
@@ -216,7 +216,7 @@ namespace SpatiaLite
         file.close();
 
         return new XmlDocument((XmlDocumentType)blob,
-                               size,
+                               (int)size,
                                XmlDocument::cleanDelete);
 
     }
